+++ packages/types-known/src/modules.ts
@@ -72,12 +72,6 @@ const typesModules: Record<string, OverrideModuleType> = {
   },
   treasury: {
     Proposal: 'TreasuryProposal'
-  },
-  xcm: {
-    Outcome: 'XcmOutcome'
-  },
-  xcmPallet: {
-    Outcome: 'XcmOutcome'
   }
 };