+++ packages/metadata/src/decorate/storage/index.ts
@@ -11,7 +11,7 @@ import { createFunction } from './createFunction';
 import { getStorage } from './getStorage';
 
 /** @internal */
-export function decorateStorage (registry: Registry, { modules }: MetadataLatest, metaVersion: number): Storage {
+export function decorateStorage (registry: Registry, { modules }: MetadataLatest): Storage {
   return modules.reduce((result: Storage, moduleMetadata): Storage => {
     if (moduleMetadata.storage.isNone) {
       return result;
@@ -31,11 +31,11 @@ export function decorateStorage (registry: Registry, { modules }: MetadataLatest
         method,
         prefix,
         section
-      }, { metaVersion });
+      }, {});
 
       return newModule;
     }, {} as ModuleStorage);
 
     return result;
-  }, { ...getStorage(registry, metaVersion) });
+  }, { ...getStorage(registry) });
 }