+++ packages/metadata/src/decorate/storage/getStorage.ts
@@ -8,12 +8,12 @@ import type { Storage } from '../types';
 import { substrate } from './substrate';
 
 /** @internal */
-export function getStorage (registry: Registry, metaVersion: number): Storage {
+export function getStorage (registry: Registry): Storage {
   return {
     substrate: Object
       .entries(substrate)
       .reduce((storage: Record<string, StorageEntry>, [key, fn]): Record<string, StorageEntry> => {
-        (storage as Record<string, unknown>)[key] = fn(registry, metaVersion);
+        (storage as Record<string, unknown>)[key] = fn(registry);
 
         return storage;
       }, {})