+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - De-alias XCmOutcome (used in multiple pallettes)
+- Fix generate of well-known storage keys (e.g. `:code`)
 
 
 ## 4.7.2 Apr 26, 2021