+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 Changes:
 
-- De-alias XCmOutcome (used in multiple pallettes)
+- De-alias xcm `Outcome` (used in multiple pallettes)
 - Fix generate of well-known storage keys (e.g. `:code`)