+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes:
 - Fix generate of well-known storage keys (e.g. `:code`)
 - Optimize metadata unknown type validation
 - Allow metadata creation for `Map` or `Record`
+- Update to latest Substrate metadata
 
 
 ## 4.7.2 Apr 26, 2021